# openapi-immix.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Immix API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "passwordConfirm": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identity": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully"
          }
        }
      }
    },
    "/api/conversations": {
      "get": {
        "summary": "Get a list of conversations",
        "responses": {
          "200": {
            "description": "A list of conversations"
          }
        }
      },
      "post": {
        "summary": "Create a new conversation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conversation created successfully"
          }
        }
      }
    },
    "/api/messages": {
      "get": {
        "summary": "Get a list of messages for a conversation",
        "parameters": [
          {
            "name": "conversation",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of messages"
          }
        }
      },
      "post": {
        "summary": "Create a new message",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "conversation": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message created successfully"
          }
        }
      }
    },
    "/api/agent/chat": {
      "post": {
        "summary": "Start a chat stream",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "conversationId": {
                    "type": "string"
                  },
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": {
                          "type": "string"
                        },
                        "content": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A stream of chat messages"
          }
        }
      }
    },
    "/api/models": {
      "get": {
        "summary": "Get a list of available models",
        "responses": {
          "200": {
            "description": "A list of models"
          }
        }
      }
    },
    "/api/models/switch": {
      "post": {
        "summary": "Switch the active model",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "path": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model switched successfully"
          }
        }
      }
    }
  }
}